backend/ai-service/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
transformers
opencv-python
easyocr
ultralytics
requests
Pillow
